Move Totalconnect coordinator to separate module
