+++ homeassistant/components/totalconnect/__init__.py
@@ -1,29 +1,20 @@
 """The totalconnect component."""
 
-from datetime import timedelta
-import logging
-
 from total_connect_client.client import TotalConnectClient
-from total_connect_client.exceptions import (
-    AuthenticationError,
-    ServiceUnavailable,
-    TotalConnectError,
-)
+from total_connect_client.exceptions import AuthenticationError
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed
 import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import AUTO_BYPASS, CONF_USERCODES, DOMAIN
+from .coordinator import TotalConnectDataUpdateCoordinator
 
 PLATFORMS = [Platform.ALARM_CONTROL_PANEL, Platform.BINARY_SENSOR, Platform.BUTTON]
 
 CONFIG_SCHEMA = cv.removed(DOMAIN, raise_if_present=False)
-SCAN_INTERVAL = timedelta(seconds=30)
-_LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -76,41 +67,3 @@ async def update_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
     client = hass.data[DOMAIN][entry.entry_id].client
     for location_id in client.locations:
         client.locations[location_id].auto_bypass_low_battery = bypass
-
-
-class TotalConnectDataUpdateCoordinator(DataUpdateCoordinator[None]):  # pylint: disable=hass-enforce-coordinator-module
-    """Class to fetch data from TotalConnect."""
-
-    config_entry: ConfigEntry
-
-    def __init__(self, hass: HomeAssistant, client: TotalConnectClient) -> None:
-        """Initialize."""
-        self.hass = hass
-        self.client = client
-        super().__init__(
-            hass, logger=_LOGGER, name=DOMAIN, update_interval=SCAN_INTERVAL
-        )
-
-    async def _async_update_data(self) -> None:
-        """Update data."""
-        await self.hass.async_add_executor_job(self.sync_update_data)
-
-    def sync_update_data(self) -> None:
-        """Fetch synchronous data from TotalConnect."""
-        try:
-            for location_id in self.client.locations:
-                self.client.locations[location_id].get_panel_meta_data()
-        except AuthenticationError as exception:
-            # should only encounter if password changes during operation
-            raise ConfigEntryAuthFailed(
-                "TotalConnect authentication failed during operation."
-            ) from exception
-        except ServiceUnavailable as exception:
-            raise UpdateFailed(
-                "Error connecting to TotalConnect or the service is unavailable. "
-                "Check https://status.resideo.com/ for outages."
-            ) from exception
-        except TotalConnectError as exception:
-            raise UpdateFailed(exception) from exception
-        except ValueError as exception:
-            raise UpdateFailed("Unknown state from TotalConnect") from exception
