+++ homeassistant/components/totalconnect/binary_sensor.py
@@ -17,8 +17,8 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import TotalConnectDataUpdateCoordinator
 from .const import DOMAIN
+from .coordinator import TotalConnectDataUpdateCoordinator
 from .entity import TotalConnectLocationEntity, TotalConnectZoneEntity
 
 LOW_BATTERY = "low_battery"
