+++ homeassistant/components/totalconnect/button.py
@@ -12,8 +12,8 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import TotalConnectDataUpdateCoordinator
 from .const import DOMAIN
+from .coordinator import TotalConnectDataUpdateCoordinator
 from .entity import TotalConnectLocationEntity, TotalConnectZoneEntity
